setup-environment-internal: allow build location to be out-of-tree
Currently, the build output is always placed under the main repo tree.
This may be undesirable on some systems which want to make use of an
extra storage drive or ramfs for building packages.

Let's add the necessary logic to handle an out-of-tree build location:
- setup-environment-internal: pass the OEROOT value into an includable
  conf file, if the BUILDDIR is absolute to prepend the OEROOT location
  and don't assume a path location of "../" when scanning for eula
  files from the BUILDDIR
- bblayers.conf: use the OEROOT conf value and remove the relative
  build/conf path assumption (../..)

Example: MACHINE=hikey DISTRO=rpb . setup-environment /out/build-rpb

Change-Id: I141229a7a589c50aa12970b3ea5edcb5a3d99827
Signed-off-by: Michael Scott <michael.scott@linaro.org>